remove editor from viewer package.json & modalStyle refactor
