+++ packages/ricos-viewer/web/package.json
@@ -73,8 +73,6 @@
     "jss-preset-default": "^10.1.1",
     "react-modal": "~3.4.1",
     "wix-rich-content-common": "7.8.0",
-    "wix-rich-content-editor": "7.8.0",
-    "wix-rich-content-editor-common": "7.8.0",
     "wix-rich-content-fullscreen": "7.8.0",
     "wix-rich-content-viewer": "7.8.0"
   },
